+++ data/update/1.0.1/pdo_mysql/2013-11.sql
@@ -1,0 +1,5 @@
+SET foreign_key_checks=0;
+
+INSERT INTO core_config_data (identifier, value) VALUES ('stream_wrapper_is_active', '1');
+
+SET foreign_key_checks=1;