+++ data/update/1.0.1/pdo_pgsql/2013-11.sql
@@ -1,0 +1,1 @@
+INSERT INTO core_config_data (identifier, value) VALUES ('stream_wrapper_is_active', '1');