+++ library/Gc/Layout/Model.php
@@ -147,6 +147,7 @@ class Model extends AbstractTable
                 $this->update($arraySave, array('id' => $this->getId()));
             }
 
+            file_put_contents($this->getFilePath(), $this->getContent());
             $this->events()->trigger(__CLASS__, 'after.save', null, array('object' => $this));
 
             return $this->getId();
@@ -172,6 +173,10 @@ class Model extends AbstractTable
                 throw new \Gc\Exception($e->getMessage(), $e->getCode(), $e);
             }
 
+            if (file_exists($this->getFilePath())) {
+                unlink($this->getFilePath());
+            }
+
             $this->events()->trigger(__CLASS__, 'after.delete', null, array('object' => $this));
             unset($this);
 
@@ -181,5 +186,16 @@ class Model extends AbstractTable
         $this->events()->trigger(__CLASS__, 'after.delete.failed', null, array('object' => $this));
 
         return false;
+    }
+
+    /**
+     * Return file path
+     *
+     * @return string
+     */
+    public function getFilePath()
+    {
+        $filename = GC_TEMPLATE_PATH . '/layout/%s.phtml';
+        return sprintf($filename, $this->getIdentifier());
     }
 }