+++ module/Application/src/Application/Controller/InstallController.php
@@ -175,6 +175,18 @@ class InstallController extends Action
             'value' => is_writable(GC_APPLICATION_PATH . '/data/cache')
         );
         $serverData[] = array(
+            'label' => '/module/Development/views/layout',
+            'value' => is_writable(GC_TEMPLATE_PATH . '/layout')
+        );
+        $serverData[] = array(
+            'label' => '/module/Development/views/view',
+            'value' => is_writable(GC_TEMPLATE_PATH . '/view')
+        );
+        $serverData[] = array(
+            'label' => '/module/Development/views/script',
+            'value' => is_writable(GC_TEMPLATE_PATH . '/script')
+        );
+        $serverData[] = array(
             'label' => basename(GC_PUBLIC_PATH) . '/media',
             'value' => File::isWritable(GC_MEDIA_PATH)
         );